Extract shadow pre-reload cache into sections/shadow-cache.js
Third of P-08's six region extractions. 28 LOC, zero cross-module references: both callers reach in via function-body imports (_consumeShadowCache from Tabs & Cards, _refreshShadowCache from SSE).
