+++ frontend/src/js/dashboard-main.js
@@ -27,6 +27,9 @@ import {
 import {
     openAboutModal, closeAboutModal, initAboutModalSideEffects,
 } from './sections/about-modal.js';
+import {
+    _refreshShadowCache, _consumeShadowCache,
+} from './sections/shadow-cache.js';
 
 /* --- In-app config editor --- */
 function _setField(form, name, value) {
@@ -799,35 +802,6 @@ async function _reloadInPlace() {
             _reloadInPlace();
         }
     }
-}
-
-/* --- Shadow pre-reload cache (Phase 4) ---
-   When staleness lands on an inactive tab, kick off a background
-   fetch of the dashboard HTML. On the user's next tab click that
-   would have triggered a live _reloadInPlace, the cached HTML is
-   applied instead, so the switch feels instant. The cache is a
-   single entry because `/` returns the whole app — any tab's fresh
-   content is already included. */
-var _shadowCache = null;  // {html, at} | {inflight: true} | null
-
-async function _refreshShadowCache() {
-    if (_shadowCache && _shadowCache.inflight) return;
-    _shadowCache = {inflight: true};
-    try {
-        var res = await fetch('/', {cache: 'no-store'});
-        if (!res.ok) { _shadowCache = null; return; }
-        var html = await res.text();
-        _shadowCache = {html: html, at: Date.now()};
-    } catch (e) {
-        _shadowCache = null;
-    }
-}
-
-function _consumeShadowCache() {
-    if (!_shadowCache || _shadowCache.inflight || !_shadowCache.html) return null;
-    var html = _shadowCache.html;
-    _shadowCache = null;
-    return html;
 }
 
 function _tabForNodeId(id) {
